+++ src/main/java/org/smoothbuild/task/base/ComputationHashes.java
@@ -23,6 +23,10 @@ public class ComputationHashes {
     return hash(3, function.hash());
   }
 
+  public static HashCode throwExceptionComputationHash() {
+    return hash(4);
+  }
+
   private static HashCode hash(int id, HashCode hash) {
     return Hash.newHasher().putInt(id).putBytes(hash.asBytes()).hash();
   }